+++ src/SqlServerTests/SqlServerTests.csproj
@@ -7,6 +7,7 @@
 
     <IsPackable>false</IsPackable>
     <IsTestProject>true</IsTestProject>
+    <UserSecretsId>5a78c96d-6df9-4362-ba25-4afceae69c52</UserSecretsId>
   </PropertyGroup>
 
   <ItemGroup>
